refactor: split running fba with plotting growth, cue, exchanges
