+++ simulations/growth_and_cue/figures/exudation_fluxes.png
