+++ simulations/growth_and_cue/figures/growth_and_cue.png
