+++ simulations/growth_and_cue/figures/uptake_fluxes.png
