+++ simulations/growth_and_cue/results/exchange_fluxes.csv
@@ -1,4 +1,4 @@
-,Acetate [e0],H2O [e0],H+ [e0],CO2 [e0],L-Leucine [e0],NH3 [e0],Phosphate [e0],O2 [e0],L-Isoleucine [e0],L-Valine [e0],3-Methyl-2-oxobutanoate [e0],Phenol [e0],D-Glucose [e0],(R)-3-Hydroxybutanoate [e0],L-Alanine [e0],L-Proline [e0],Nitrate [e0],Glycerol-3-phosphate [e0],Sulfite [e0],"(S)-2,3-dihydroxypropane-1-sulfonate [e0]",L-Tyrosine [e0],L-Glutamate [e0],L-Arginine [e0],L-Lysine [e0],Galactose [e0],L-Aspartate [e0],Glycine [e0],Other
+,Acetate,H2O,H+,CO2,L-Leucine,NH3,Phosphate,O2,L-Isoleucine,L-Valine,3-Methyl-2-oxobutanoate,Phenol,D-Glucose,(R)-3-Hydroxybutanoate,L-Alanine,L-Proline,Nitrate,Glycerol-3-phosphate,Sulfite,"(S)-2,3-dihydroxypropane-1-sulfonate",L-Tyrosine,L-Glutamate,L-Arginine,L-Lysine,Galactose,L-Aspartate,Glycine,Other
 Glycerol-3-phosphate,0.0,37.70786591361484,-9.82850350769882,25.45456807048622,0.0,0.0,19.34460294799136,-15.974853161083985,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,-8.324853442599352,-20.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,-0.23454863524757474
 Glucose,2.555078385032359,41.8514353178502,8.922648018331222,22.64180264971312,0.0,-7.771221740376767,-0.611810869016085,-20.0,0.0,0.0,0.0,0.0,-10.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,-0.21895033539984696
 Galactose,3.0233230064349046,41.26491407783888,9.219978702829536,22.57089319486795,0.0,-7.562631936308094,-0.5953890612799481,-20.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,-10.0,0.0,0.0,-0.2130734206639684
