+++ simulations/growth_and_cue/results/substrate_panel.csv
@@ -1,0 +1,20 @@
+name,name_in_model,met_id,exchange_id,media_key,n_c,entry_point
+Leucine,L-Leucine,cpd00107,EX_cpd00107_e0,promm_no_c,6,Acetyl-CoA
+Isoleucine,L-Isoleucine,cpd00322,EX_cpd00322_e0,promm_no_c,6,TCA — Succinyl-CoA
+Valine,L-Valine,cpd00156,EX_cpd00156_e0,promm_no_c,5,TCA — Succinyl-CoA
+3-methyl-2-oxobutanoic acid,3-Methyl-2-oxobutanoate,cpd00123,EX_cpd00123_e0,promm_no_c,5,TCA — Succinyl-CoA
+Phenol,Phenol,cpd00127,EX_cpd00127_e0,swm,6,Aromatic catabolism
+Glucose,D-Glucose,cpd00027,EX_cpd00027_e0,mbm,6,Glycolysis
+Acetate,Acetate,cpd00029,EX_cpd00029_e0,mbm,2,Acetyl-CoA
+3-hydroxybutyrate,(R)-3-Hydroxybutanoate,cpd00797,EX_cpd00797_e0,mbm,4,Acetyl-CoA
+Alanine,L-Alanine,cpd00035,EX_cpd00035_e0,l1,3,Pyruvate
+Proline,L-Proline,cpd00129,EX_cpd00129_e0,l1,5,TCA — α-KG
+Glycerol-3-phosphate,Glycerol-3-phosphate,cpd00080,EX_cpd00080_e0,l1,3,Glycolysis
+DHPS (dihydroxypropanesulfonate),"(S)-2,3-dihydroxypropane-1-sulfonate",cpd23538,EX_cpd23538_e0,l1,3,Pyruvate
+Tyrosine,L-Tyrosine,cpd00069,EX_cpd00069_e0,l1,9,Aromatic catabolism
+Glutamate,L-Glutamate,cpd00023,EX_cpd00023_e0,l1,5,TCA — α-KG
+Arginine,L-Arginine,cpd00051,EX_cpd00051_e0,marine_broth_wo_yeast_and_peptone,6,TCA — α-KG
+Lysine,L-Lysine,cpd00039,EX_cpd00039_e0,marine_broth_wo_yeast_and_peptone,6,Acetyl-CoA
+Galactose,Galactose,cpd00108,EX_cpd00108_e0,mbm,6,Glycolysis
+Aspartate,L-Aspartate,cpd00041,EX_cpd00041_e0,l1,4,TCA — OAA
+Glycine,Glycine,cpd00033,EX_cpd00033_e0,l1,2,Pyruvate
