+++ simulations/growth_and_cue/run_growth_cue.py
@@ -15,39 +15,17 @@ import warnings
 from pathlib import Path
 from typing import Optional
 import pickle as pkl
-import sys
 
 import cobra
 import cobra.flux_analysis
 from gem_utilities import media as media_utils
-import matplotlib
-
-matplotlib.use("Agg")
-import matplotlib.pyplot as plt
-import matplotlib.patches as mpatches
-from matplotlib.lines import Line2D
-import numpy as np
 import pandas as pd
-from scipy import stats
-
-matplotlib.rcParams.update(
-    {
-        "font.size": 11,
-        "axes.linewidth": 0.8,
-        "axes.spines.top": False,
-        "axes.spines.right": False,
-    }
-)
 
 FILE_PATH = Path(__file__).resolve().parent
 REPO_ROOT = FILE_PATH.parents[1]
 TEST_FILE_DIR = REPO_ROOT / "test" / "test_files"
 OUT_PATH = FILE_PATH / "results"
 OUT_PATH.mkdir(exist_ok=True)
-
-# Import plot_styles.py from the root of the repo
-sys.path.append(str(REPO_ROOT))
-from plot_styles import summer_colors
 
 TOTAL_UPTAKE = 60  # mmol C / gDW / hr
 BIOMASS_RXN = "bio1_biomass"
@@ -56,35 +34,6 @@ CO2_EX_RXN = "EX_cpd00011_e0"
 # Exchange metabolites whose max |flux| across substrates is below this are
 # lumped into a single grey "Other" segment (keeps the trace-ion colours out).
 EX_FLUX_THRESHOLD = 1.0  # mmol / gDW / hr
-
-# Color palette for exchange metabolites
-# The "Summer" color palette with a few extra colors to avoid repeats
-# Remove colors that I hard set below
-EX_PALETTE = [
-    v
-    for k, v in summer_colors.items()
-    if k not in {"yellow", "light_blue", "dark_tan", "light_tan"}
-]
-EX_PALETTE += [
-    "#9B6A9E",  # mauve
-    "#3F6F8C",  # steel blue
-    "#C98B8B",  # dusty rose
-    "#8C4F6B",  # plum
-    "#6E7B8B",  # slate
-    "#BFB13A",  # brass
-]
-# Define an eye-catching color for ammonium, since that is the primary result
-NH3_COLOR = summer_colors["yellow"]
-# Define muted colors for metabolites that are unimportant
-H_COLOR = summer_colors["light_blue"]
-CO2_COLOR = summer_colors["dark_tan"]
-H2O_COLOR = summer_colors["light_tan"]
-# Define a single color for the carbon source
-# So there aren't unique colros for each bar
-# OK to be the same color as CO2, since the carbon source is ony ever taken up
-# And CO2 is only ever released
-CARBON_SOURCE_COLOR = summer_colors["dark_tan"]  # all substrate carbon sources
-OTHER_COLOR = "#bdbdbd"  # grey — collapsed trace metabolites
 
 # Entry point into central metabolism = the first central-metabolism
 # intermediate the substrate's catabolism produces.
@@ -110,17 +59,32 @@ ENTRY_POINT = {
     "cpd00127": "Aromatic catabolism",  # Phenol       → succinyl-CoA + AcCoA
 }
 
-# SMF-figure palette, extended with two harmonising tones for the 7 categories.
-ENTRY_POINT_COLORS = {
-    "Glycolysis": summer_colors["light_blue"],
-    "Pyruvate": summer_colors["teal"],
-    "Acetyl-CoA": summer_colors["yellow"],
-    "TCA — α-KG": summer_colors["pink"],
-    "TCA — OAA": summer_colors["dark_pink"],
-    "TCA — Succinyl-CoA": summer_colors["green"],
-    "Aromatic catabolism": "#9B6A9E",  # muted mauve (added)
-}
-ENTRY_POINT_ORDER = list(ENTRY_POINT_COLORS)
+
+def main():
+    print("Loading model...")
+    model = cobra.io.read_sbml_model(REPO_ROOT / "model.xml")
+
+    print("Loading media definitions...")
+    with open(TEST_FILE_DIR / "media" / "media_definitions.pkl", "rb") as f:
+        media_defs = pkl.load(f)
+
+    print("\nBuilding substrate panel...")
+    substrate_df = load_substrates(model, media_defs)
+
+    # Save the substrate df
+    substrate_df.to_csv(OUT_PATH / "substrate_panel.csv", index=False)
+
+    print("\nRunning pFBA simulations...")
+    summary_df, ex_records = run_pfba(model, media_defs, substrate_df)
+    print(f"\nSuccessful: {len(summary_df)}/{len(substrate_df)} substrates")
+
+    # Save the results
+    summary_df.to_csv(OUT_PATH / "growth_and_cue.csv")
+
+    # Extract the exchange fluxes
+    order = summary_df.sort_values("growth_rate", ascending=False).index.tolist()
+    ex_df = build_exchange_df(model, ex_records, order, EX_FLUX_THRESHOLD)
+    ex_df.to_csv(OUT_PATH / "exchange_fluxes.csv")
 
 
 def count_carbons(formula: str) -> Optional[int]:
@@ -196,11 +160,15 @@ def load_substrates(model: cobra.Model, media_defs: dict) -> pd.DataFrame:
 
         try:
             n_c = count_carbons(model.metabolites.get_by_id(f"{met_id}_e0").formula)
+            model_name = model.metabolites.get_by_id(f"{met_id}_e0").name[
+                :-5
+            ]  # drop " [e0]" suffix
         except KeyError:
             n_c = None
             for met in model.metabolites:
                 if met.id.startswith(met_id):
                     n_c = count_carbons(met.formula)
+                    model_name = met.name
                     break
         if not n_c:
             print(f"  SKIP (can't determine n_c): {c_source}")
@@ -209,6 +177,7 @@ def load_substrates(model: cobra.Model, media_defs: dict) -> pd.DataFrame:
         records.append(
             {
                 "name": c_source,
+                "name_in_model": model_name,
                 "met_id": met_id,
                 "exchange_id": ex_id,
                 "media_key": media_key,
@@ -268,132 +237,6 @@ def run_pfba(model, media_defs, substrate_df) -> tuple:
     return pd.DataFrame(rows).set_index("substrate"), ex_records
 
 
-def plot_growth_cue(summary_df: pd.DataFrame, out_path: Path) -> None:
-    df = summary_df.sort_values("growth_rate", ascending=False).reset_index()
-    colors = [ENTRY_POINT_COLORS.get(c, "#9E9E9E") for c in df["entry_point"]]
-    x = np.arange(len(df))
-
-    fig, (ax_g, ax_c) = plt.subplots(2, 1, figsize=(13, 8), sharex=True)
-
-    ax_g.bar(x, df["growth_rate"], color=colors, edgecolor="white", linewidth=0.6)
-    ax_g.set_ylabel("Growth rate (h⁻¹)")
-    ax_g.set_title(
-        "MIT1002 growth rate and carbon-use efficiency across substrates",
-        fontsize=12,
-        pad=8,
-    )
-    ax_g.margins(x=0.01)
-
-    ax_c.bar(x, df["cue"], color=colors, edgecolor="white", linewidth=0.6)
-    ax_c.set_ylabel("Carbon-use efficiency")
-    ax_c.set_ylim(0, 1)
-    ax_c.set_xticks(x)
-    ax_c.set_xticklabels(df["substrate"], rotation=45, ha="right", fontsize=9)
-
-    handles = [
-        mpatches.Patch(facecolor=ENTRY_POINT_COLORS[lbl], edgecolor="white", label=lbl)
-        for lbl in ENTRY_POINT_ORDER
-        if lbl in df["entry_point"].values
-    ]
-    ax_g.legend(
-        handles=handles,
-        title="Entry point into\ncentral metabolism",
-        frameon=False,
-        bbox_to_anchor=(1.01, 1),
-        loc="upper left",
-        fontsize=9,
-        title_fontsize=9.5,
-    )
-
-    fig.tight_layout()
-    fig.savefig(out_path, dpi=300, bbox_inches="tight")
-    plt.close(fig)
-    print(f"  Saved: {out_path.name}")
-
-
-def plot_growth_cue_correlation(summary_df: pd.DataFrame, out_path: Path) -> None:
-    df = summary_df.reset_index()
-    colors = [ENTRY_POINT_COLORS.get(c, "#9E9E9E") for c in df["entry_point"]]
-    x = df["growth_rate"].values.astype(float)
-    y = df["cue"].values.astype(float)
-
-    r, p = stats.pearsonr(x, y)
-    rho, p_s = stats.spearmanr(x, y)
-
-    fig, ax = plt.subplots(figsize=(8.5, 6.5))
-    ax.scatter(x, y, c=colors, s=120, edgecolor="white", linewidth=0.6, zorder=3)
-
-    # Least-squares fit line
-    slope, intercept = np.polyfit(x, y, 1)
-    xs = np.linspace(x.min(), x.max(), 100)
-    ax.plot(xs, slope * xs + intercept, color="#555555", ls="--", lw=1.2, zorder=2)
-
-    # Substrate labels. Default is offset to the right of the point; a few
-    # near-coincident points get a manual override so labels don't overlap.
-    label_offsets = {  # substrate: (dx, dy, ha, va)
-        "Galactose": (0, 9, "center", "bottom"),  # sits almost on Glucose
-    }
-    for _, row in df.iterrows():
-        dx, dy, ha, va = label_offsets.get(row["substrate"], (5, 0, "left", "center"))
-        ax.annotate(
-            row["substrate"],
-            (row["growth_rate"], row["cue"]),
-            xytext=(dx, dy),
-            textcoords="offset points",
-            fontsize=7,
-            ha=ha,
-            va=va,
-            color="#555555",
-        )
-
-    ax.set_xlabel("Growth rate (h⁻¹)")
-    ax.set_ylabel("Carbon-use efficiency")
-    ax.set_title(
-        "Growth rate vs. carbon-use efficiency across substrates", fontsize=12, pad=8
-    )
-
-    # Correlation metrics
-    txt = (
-        f"Pearson  r = {r:.2f}  (p = {p:.2g})\n"
-        f"Spearman ρ = {rho:.2f}  (p = {p_s:.2g})"
-    )
-    ax.text(
-        0.025,
-        0.025,
-        txt,
-        transform=ax.transAxes,
-        fontsize=9,
-        va="bottom",
-        ha="left",
-        linespacing=1.4,
-        bbox=dict(boxstyle="round,pad=0.4", fc="white", ec="#cccccc", alpha=0.9),
-    )
-
-    # Legend: entry-point colors + the fit line
-    handles = [
-        mpatches.Patch(facecolor=ENTRY_POINT_COLORS[lbl], edgecolor="white", label=lbl)
-        for lbl in ENTRY_POINT_ORDER
-        if lbl in df["entry_point"].values
-    ]
-    handles.append(
-        Line2D([0], [0], color="#555555", ls="--", lw=1.2, label="linear fit")
-    )
-    ax.legend(
-        handles=handles,
-        title="Entry point into\ncentral metabolism",
-        frameon=False,
-        bbox_to_anchor=(1.01, 1),
-        loc="upper left",
-        fontsize=9,
-        title_fontsize=9.5,
-    )
-
-    fig.tight_layout()
-    fig.savefig(out_path, dpi=300, bbox_inches="tight")
-    plt.close(fig)
-    print(f"  Saved: {out_path.name}  (Pearson r={r:.2f}, p={p:.2g})")
-
-
 def build_exchange_df(model, ex_records, substrate_order, threshold):
     """{substrate: {ex_rxn: flux}} -> DataFrame (substrate x metabolite name).
 
@@ -401,11 +244,18 @@ def build_exchange_df(model, ex_records, substrate_order, threshold):
     Trace metabolites (max |flux| < threshold) are collapsed into 'Other'.
     """
     df = pd.DataFrame(ex_records).T.reindex(substrate_order).fillna(0.0)
+    # Rename the column with the metabolite name instead of the reaction ID
     rename = {}
     for rid in df.columns:
         met = next(iter(model.reactions.get_by_id(rid).metabolites))
-        rename[rid] = met.name
+        met_name = met.name
+        # If the name ends with " [e0]", drop that suffix for cleaner labels (extracellular met)
+        suffix = " [e0]"
+        if met_name.endswith(suffix):
+            met_name = met_name[: -len(suffix)]
+        rename[rid] = met_name
     df = df.rename(columns=rename)
+
     # collapse duplicate metabolite-name columns if any arise (version-safe)
     if df.columns.duplicated().any():
         df = df.T.groupby(level=0).sum().T
@@ -418,167 +268,6 @@ def build_exchange_df(model, ex_records, substrate_order, threshold):
             out["Other"] = df[small].sum(axis=1)
         df = out
     return df
-
-
-def _merge_carbon_sources(df, carbon_source_names):
-    """Collapse all substrate-carbon-source columns into one 'Carbon source'."""
-    cs = [c for c in df.columns if c in carbon_source_names]
-    rest = [c for c in df.columns if c not in carbon_source_names]
-    out = df[rest].copy()
-    if cs:
-        out["Carbon source"] = df[cs].sum(axis=1)
-    return out
-
-
-def _ordered_cols(df, pin_first=None, pin_last=None):
-    """Column order: pinned-first, then regular by descending magnitude, pinned-last."""
-    skip = {pin_first, pin_last}
-    middle = sorted(
-        (c for c in df.columns if c not in skip),
-        key=lambda c: df[c].abs().sum(),
-        reverse=True,
-    )
-    head = [pin_first] if pin_first in df.columns else []
-    tail = [pin_last] if pin_last in df.columns else []
-    return head + middle + tail
-
-
-def _stacked_bar(df, colors, title, ylabel, out_path):
-    fig, ax = plt.subplots(figsize=(13, 7))
-    x = np.arange(len(df.index))
-    bottom = np.zeros(len(df))
-    for col in df.columns:
-        vals = df[col].values
-        ax.bar(
-            x,
-            vals,
-            bottom=bottom,
-            color=colors[col],
-            label=col,
-            edgecolor="white",
-            linewidth=0.3,
-        )
-        bottom += vals
-    ax.axhline(0, color="black", lw=0.6)
-    ax.set_xticks(x)
-    ax.set_xticklabels(df.index, rotation=45, ha="right", fontsize=8)
-    ax.set_ylabel(ylabel)
-    ax.set_title(title, fontsize=12, pad=8)
-    ax.margins(x=0.01)
-    ax.legend(
-        title="Metabolite",
-        bbox_to_anchor=(1.01, 1),
-        loc="upper left",
-        fontsize=7,
-        title_fontsize=8.5,
-        frameon=False,
-    )
-    fig.tight_layout()
-    fig.savefig(out_path, dpi=300, bbox_inches="tight")
-    plt.close(fig)
-    print(f"  Saved: {out_path.name}")
-
-
-def plot_exchange_stacks(ex_df, carbon_source_names, out_dir):
-    """Two stacked-bar charts, both drawn upward, with shared metabolite colours.
-
-    Uptake fluxes are negated so they read as positive bars. All substrate
-    carbon sources are merged into one 'Carbon source' segment in the uptake
-    chart (each appears on a single bar). Byproducts/co-substrates keep
-    consistent colours across both charts.
-    """
-    # Separate uptake and exudation
-    uptake = -ex_df.clip(upper=0)  # negative fluxes -> positive uptake
-    exud = ex_df.clip(lower=0)  # positive fluxes -> exudation
-
-    # Drop zero-only columns (e.g. if no exudation of a metabolite occurs)
-    uptake = uptake.loc[:, (uptake != 0).any(axis=0)]
-    exud = exud.loc[:, (exud != 0).any(axis=0)]
-
-    # Merge all substrate carbon sources into one column in the uptake chart
-    # So they don't have unique colours and the chart is less busy
-    uptake = _merge_carbon_sources(uptake, carbon_source_names)
-
-    # Drop H2O from the exudation chart
-    exud = exud.drop(columns=["H2O"], errors="ignore")
-
-    # Shared colour map over the "regular" metabolites (everything except the
-    # fixed-colour Carbon source / Other), ordered by total magnitude so shared
-    # metabolites get the same colour in both charts.
-    special = {"H2O", "H+", "CO2", "NH3", "Carbon source", "Other"}
-    regular = (set(uptake.columns) | set(exud.columns)) - special
-
-    def total_mag(c):
-        m = 0.0
-        if c in uptake.columns:
-            m += uptake[c].abs().sum()
-        if c in exud.columns:
-            m += exud[c].abs().sum()
-        return m
-
-    ordered_reg = sorted(regular, key=total_mag, reverse=True)
-    colors = {c: EX_PALETTE[i % len(EX_PALETTE)] for i, c in enumerate(ordered_reg)}
-    # Define the special colors
-    colors["H+"] = H_COLOR
-    colors["H2O"] = H2O_COLOR
-    colors["CO2"] = CO2_COLOR
-    colors["NH3"] = NH3_COLOR
-    colors["Carbon source"] = CARBON_SOURCE_COLOR
-    colors["Other"] = OTHER_COLOR
-
-    up_cols = _ordered_cols(uptake, pin_first="Carbon source", pin_last="Other")
-    ex_cols = _ordered_cols(exud, pin_last="Other")
-    _stacked_bar(
-        uptake[up_cols],
-        colors,
-        "MIT1002 uptake fluxes across substrates",
-        "Uptake flux (mmol gDW⁻¹ h⁻¹)",
-        out_dir / "uptake_fluxes.png",
-    )
-    _stacked_bar(
-        exud[ex_cols],
-        colors,
-        "MIT1002 exudation fluxes across substrates",
-        "Exudation flux (mmol gDW⁻¹ h⁻¹)",
-        out_dir / "exudation_fluxes.png",
-    )
-
-
-def main():
-    print("Loading model...")
-    model = cobra.io.read_sbml_model(REPO_ROOT / "model.xml")
-
-    print("Loading media definitions...")
-    with open(TEST_FILE_DIR / "media" / "media_definitions.pkl", "rb") as f:
-        media_defs = pkl.load(f)
-
-    print("\nBuilding substrate panel...")
-    substrate_df = load_substrates(model, media_defs)
-
-    print("\nRunning pFBA simulations...")
-    summary_df, ex_records = run_pfba(model, media_defs, substrate_df)
-    print(f"\nSuccessful: {len(summary_df)}/{len(substrate_df)} substrates")
-
-    summary_df.to_csv(OUT_PATH / "growth_and_cue.csv")
-    print("\nPlotting growth rate and CUE...")
-    plot_growth_cue(summary_df, OUT_PATH / "growth_and_cue.png")
-    plot_growth_cue_correlation(summary_df, OUT_PATH / "growth_vs_cue_scatter.png")
-
-    # Exchange (uptake/exudation) stacked bars, substrates ordered by growth rate
-    print("\nPlotting exchange fluxes...")
-    order = summary_df.sort_values("growth_rate", ascending=False).index.tolist()
-    ex_df = build_exchange_df(model, ex_records, order, EX_FLUX_THRESHOLD)
-    ex_df.to_csv(OUT_PATH / "exchange_fluxes.csv")
-    # Drop the " [e0]" suffix from the metabolite names for cleaner labels
-    ex_df = ex_df.rename(columns=lambda c: c.replace(" [e0]", ""))
-    # Metabolite names that are substrate carbon sources (merged in uptake chart)
-    # Remove the " [e0]" suffix from the names with "[:-5]"
-    carbon_source_names = {
-        next(iter(model.reactions.get_by_id(r["exchange_id"]).metabolites)).name[:-5]
-        for _, r in substrate_df.iterrows()
-    }
-    plot_exchange_stacks(ex_df, carbon_source_names, OUT_PATH)
-    print(f"\nAll results saved to: {OUT_PATH}")
 
 
 if __name__ == "__main__":
